requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.3.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
Flask==3.1.0
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflect==7.5.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
Mako==1.3.10
MarkupSafe==2.1.5
more-itertools==10.7.0
oauthlib==3.2.2
openai==1.77.0
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyMuPDF==1.25.5
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==7.4.0
python-dotenv==1.1.0
python-http-client==3.3.7
pytz==2023.3
rapidfuzz==3.9.7
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.20.0
sendgrid==6.12.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typeguard==4.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
URLObject==2.4.3
webdriver-manager==4.0.1
Werkzeug==3.1.3
word2number==1.1
wsproto==1.2.0
WTForms==3.1.1
